複数画面で利用できるようにリファクタリング
