+++ samples/PersistSample/persist_sample/lib/todo.dart
@@ -1,0 +1,9 @@
+class ToDo {
+  int id;
+  String title;
+
+  ToDo(int id, String title) {
+    this.id = id;
+    this.title = title;
+  }
+}
